Widen format_delta so large intervals are not truncated
format_delta cast a 64-bit time_t down to long, so on a 32-bit platform
with 64-bit time_t an interval beyond about 68 years wrapped negative and
printed 0 seconds. Compute in long long instead.

+++ notify.c
@@ -173,7 +173,7 @@ format_timestamp(time_t t, char *buf, size_t len)
 void
 format_delta(time_t delta, char *buf, size_t len)
 {
-	long d = (long)delta;
+	long long d = (long long)delta;
 
 	if (d < 0) {
 		snprintf(buf, len, "0 seconds");
@@ -181,15 +181,15 @@ format_delta(time_t delta, char *buf, size_t len)
 	}
 
 	if (d < 60)
-		snprintf(buf, len, "%ld second%s", d, d == 1 ? "" : "s");
+		snprintf(buf, len, "%lld second%s", d, d == 1 ? "" : "s");
 	else if (d < 3600)
-		snprintf(buf, len, "%ld minute%s", d / 60,
+		snprintf(buf, len, "%lld minute%s", d / 60,
 		    d / 60 == 1 ? "" : "s");
 	else if (d < 86400)
-		snprintf(buf, len, "%ld hour%s", d / 3600,
+		snprintf(buf, len, "%lld hour%s", d / 3600,
 		    d / 3600 == 1 ? "" : "s");
 	else
-		snprintf(buf, len, "%ld day%s", d / 86400,
+		snprintf(buf, len, "%lld day%s", d / 86400,
 		    d / 86400 == 1 ? "" : "s");
 }
